backend/services/messaging-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pyjwt==2.8.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
boto3==1.34.0
botocore==1.34.0
